set crashpad executable permissions (mac CI config)
* fix crashpad errors in the test logs
* ignore removing the Crashpad folder

test_osn_global: get memoryUsage value (fix typo?)
unique_object_manager: fix for possible race condition
osn-video.cpp: check for canvas* & fix xcode warning blog() string literal
video.cpp: setflag to indicate video device was invalidated
test_osn_global -> use synchronous func instead of lamda

+++ tests/osn-tests/util/general.ts
@@ -80,7 +80,8 @@ export function deleteConfigFiles(): void {
         let currentFile: string;
         let files = fs.readdirSync(configFolderPath);
         files.forEach(file => {
-            if (file !== 'node-obs') {
+            // Ignore mac Crashpad and node-obs/logs folders.
+            if (file !== 'node-obs' && file !== 'Crashpad') {
                 currentFile = file;
                 try {
                     fs.unlinkSync(path.join(configFolderPath, file));
